Refactor logo services

+++ src/main/java/de/innovationhub/prox/companyprofileservice/application/service/company/CompanyLogoServiceImpl.java
@@ -1,5 +1,6 @@
 package de.innovationhub.prox.companyprofileservice.application.service.company;
 
+import de.innovationhub.prox.companyprofileservice.domain.company.Company;
 import de.innovationhub.prox.companyprofileservice.domain.company.CompanyLogo;
 import de.innovationhub.prox.companyprofileservice.domain.company.CompanyLogoRepository;
 import de.innovationhub.prox.companyprofileservice.domain.company.CompanyLogoStore;
@@ -23,62 +24,51 @@ public class CompanyLogoServiceImpl implements CompanyLogoService {
 
   private final CompanyLogoRepository companyLogoRepository;
   private final CompanyLogoStore companyLogoStore;
+  private final CompanyService companyService;
   private final ModelMapper modelMapper;
   private final TikaConfig tikaConfig;
 
   @Autowired
-  public CompanyLogoServiceImpl(CompanyLogoRepository companyLogoRepository, CompanyLogoStore companyLogoStore) {
+  public CompanyLogoServiceImpl(CompanyLogoRepository companyLogoRepository, CompanyLogoStore companyLogoStore, CompanyService companyService) {
     this.companyLogoRepository = companyLogoRepository;
     this.companyLogoStore = companyLogoStore;
+    this.companyService = companyService;
     this.modelMapper = new ModelMapper();
     this.tikaConfig = TikaConfig.getDefaultConfig();
   }
 
-  @Override
-  public Iterable<CompanyLogo> getAll() {
-    return this.companyLogoRepository.findAll();
-  }
-
-  @Override
-  public Optional<CompanyLogo> getById(UUID uuid) {
+  private Optional<CompanyLogo> getById(UUID uuid) {
     return this.companyLogoRepository.findById(uuid);
   }
 
-  @Override
-  public CompanyLogo save(CompanyLogo entity) {
+  private CompanyLogo save(CompanyLogo entity) {
     return this.companyLogoRepository.save(entity);
   }
 
-  @Override
-  public CompanyLogo update(UUID uuid, CompanyLogo entity) {
-    return this.getById(uuid)
-        .map(
-            c -> {
-              modelMapper.map(entity, c);
-              return c;
-            })
-        .map(companyLogoRepository::save)
-        .orElseThrow(RuntimeException::new);
-  }
-
-  @Override
-  public void deleteById(UUID uuid) {
+  private void deleteById(UUID uuid) {
     this.companyLogoRepository.deleteById(uuid);
   }
 
-
   @Override
-  public Optional<InputStream> getCompanyLogoAsStream(CompanyLogo companyLogo) {
-    if(companyLogo != null) {
-      InputStream is = companyLogoStore.getContent(companyLogo);
-      return Optional.of(is);
-    }
-    return Optional.empty();
+  public Optional<CompanyLogo> getCompanyLogo(UUID companyId) {
+    return this.companyService.getById(companyId)
+        .map(Company::getCompanyLogo);
   }
 
   @Override
-  public Optional<CompanyLogo> setCompanyLogo(CompanyLogo companyLogo, InputStream inputStream)
+  public Optional<InputStream> getCompanyLogoAsStream(CompanyLogo companyLogo) {
+    return Optional.of(this.companyLogoStore.getContent(companyLogo));
+  }
+
+  @Override
+  public Optional<CompanyLogo> setCompanyLogo(UUID companyId, InputStream inputStream)
       throws IOException {
+    var optCompany = this.companyService.getById(companyId);
+    if(optCompany.isEmpty()) {
+      return Optional.empty();
+    }
+    var company = optCompany.get();
+    var companyLogo = company.getCompanyLogo();
     if(companyLogo == null) {
       companyLogo = new CompanyLogo();
     }
@@ -94,13 +84,24 @@ public class CompanyLogoServiceImpl implements CompanyLogoService {
     CompanyLogo companyLogo1 = companyLogoStore.setContent(companyLogo, new ByteArrayInputStream(bytes));
     companyLogo = this.save(companyLogo1);
 
+    company.setCompanyLogo(companyLogo);
+    this.companyService.save(company);
+
     return Optional.of(companyLogo);
   }
 
   @Override
-  public Optional<CompanyLogo> deleteCompanyLogo(CompanyLogo companyLogo) {
+  public Optional<CompanyLogo> deleteCompanyLogo(UUID companyId) {
+    var optCompany = this.companyService.getById(companyId);
+    if(optCompany.isEmpty()) {
+      return Optional.empty();
+    }
+    var company = optCompany.get();
+    var companyLogo = company.getCompanyLogo();
     if(companyLogo != null) {
       var cl = companyLogoStore.unsetContent(companyLogo);
+      company.setCompanyLogo(null);
+      this.companyService.save(company);
       this.deleteById(cl.getId());
       return Optional.of(cl);
     }
